.ipynb_checkpoints/requirements-checkpoint.txt
flask
azure-ai-textanalytics
azure-core